backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
feedparser==6.0.11
httpx==0.27.0
beautifulsoup4==4.12.3
anthropic==0.28.0
yfinance==0.2.40
apscheduler==3.10.4
python-multipart==0.0.9
aiosqlite==0.20.0
pytz==2024.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
